pyproject.toml
[project]
name = "birthday-emailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-apscheduler>=1.13.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
